refactor(providers): split manager internals
