+++ src/lib/__tests__/architecture-boundaries.test.ts
@@ -358,6 +358,25 @@ describe("architecture import boundaries", () => {
     expect(offenders).toEqual([])
   })
 
+  it("keeps ProviderManager as a facade over config and mapping storage", () => {
+    const source = readFileSync(
+      join(sourceRoot, "lib/providers/manager.ts"),
+      "utf8"
+    )
+    const forbidden = referencedModules(source).filter((module) =>
+      [
+        "@/lib/plasmo-global-storage",
+        "@/lib/constants",
+        "./provider-config-schema"
+      ].includes(module)
+    )
+
+    expect(forbidden).toEqual([])
+    expect(withoutComments(source)).not.toMatch(
+      /ProviderStorageKey|LEGACY_STORAGE_KEYS|MODEL_MAPPINGS/
+    )
+  })
+
   /**
    * The observer registry is in-memory delivery state and nothing else.
    *
